requirements.txt
streamlit>=1.28.0
python-dotenv>=1.0.0
pandas>=2.0.0
requests>=2.31.0
langchain>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
langchain-core>=0.3.0
langchain-text-splitters>=0.3.0
faiss-cpu>=1.8.0
numpy>=1.24.0
plotly>=5.17.0
yfinance>=0.2.28
wbgapi>=1.0.12
scikit-learn>=1.3.0
python-dateutil>=2.8.2
pytz>=2023.3
alpha-vantage>=2.3.1
gunicorn>=21.2.0
duckduckgo-search>=6.1.0